requirements.txt
asn1crypto==0.24.0
bcrypt==3.1.6
certifi==2019.3.9
cffi==1.12.2
chardet==3.0.4
cryptography==2.6.1
Django==2.1.7
django-cors-headers==2.5.2
djangorestframework==3.9.2
idna==2.8
ndg-httpsclient==0.5.1
Pillow==5.4.1
pusher==2.1.3
pyasn1==0.4.5
pycparser==2.19
PyNaCl==1.3.0
pyOpenSSL==19.0.0
pytz==2018.9
requests==2.21.0
six==1.12.0
urllib3==1.24.1
